+++ superset-frontend/src/dashboard/components/BuilderComponentPane/ComponentsTab/ComponentsTab.tsx
@@ -1,0 +1,20 @@
+import React, { useState } from 'react';
+
+import { Controls } from './Controls';
+import { CustomMarkdowns } from './CustomComponents';
+import { LegacyComponents } from './LegacyComponents';
+
+export function ComponentsTab() {
+  const [searchTerm, setSearchTerm] = useState('');
+
+  return (
+    <div className="custom-components-container">
+      <Controls searchTerm={searchTerm} setSearchTerm={setSearchTerm} />
+
+      <div className="custom-components-content sidepane-padding">
+        <CustomMarkdowns searchTerm={searchTerm} />
+        <LegacyComponents />
+      </div>
+    </div>
+  );
+}
